requirements.txt
Flask
boto3
prometheus-client
psutil
gunicorn
werkzeug
python-dotenv==1.0.0